Add style and content to third part
Did not use flex display this time around as the normal CSS properties
suffice, merely did text-align and using of padding to achieve the
results in the picture. Also, the quote is Stalin's but I adapted it to
cat form. This is just to poke fun at a cat saying such sophisticated
things as someone may or may not be bad.