Add style and content to final quarter
Used 3 different containers for this, one is the big container for the
quarter. The next one is for the blue rectangle. Then, to align the
button to the 2 text divs, the text divs had to be wrapped for the
format to take place. There is surely more concise ways to do it but
this is how I did it. It is not optimised for any screen size other than
the fullscreen.

The style.css sheet is also a mess. I think some properties can be
brought into an overlapping one for less repeated code and better
presentation.